refactor: modularize manifest
